+++ app/Http/Controllers/AdminApi/StatusController.php
@@ -2,6 +2,8 @@
 
 namespace BristolSU\Module\Typeform\Http\Controllers\AdminApi;
 
+use BristolSU\Module\Typeform\Events\ResponseApproved;
+use BristolSU\Module\Typeform\Events\ResponseRejected;
 use BristolSU\Module\Typeform\Http\Controllers\Controller;
 use BristolSU\Module\Typeform\Models\Response;
 use BristolSU\Support\Activity\Activity;
@@ -15,6 +17,10 @@ class StatusController extends Controller
         $this->authorize('admin.approve');
         $response->approved = true;
         $response->save();
+        $response->refresh();
+        
+        event(new ResponseApproved($response));
+        
         return $response;
     }
 
@@ -23,6 +29,10 @@ class StatusController extends Controller
         $this->authorize('admin.approve');
         $response->approved = false;
         $response->save();
+        $response->refresh();
+        
+        event(new ResponseRejected($response));
+
         return $response;
     }